==> 0_libraries.R <==
# Load libraries
pacman::p_load(
  readr,
  qdap,
  stm,
  tidytext,
  parallel,
  udpipe,
  rstudioapi,
  textstem,
  tm,
  ggwordcloud,
  gridExtra,
  stminsights,
  egg,
  ggplotify,
  lemon,
  mobr,
  future,
  tidytext,
  ggpubr
)

require(tidyverse)
